+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - Update Camel Quarkus Catalog from 3.6.0 to 3.7.0
 - Provide entry in `File > New file...` menu to create a Camel file
+- Provide command to create a Custom Resource Pipe with YAML DSL
 
 ## 0.15.0